App now support Landscape
